Rename wigners.py to wigner.py
